+++ data-raw/make_gss_dict.R
@@ -1,5 +1,5 @@
 ## 2.
-## Run make_gss_doc.R first.
+## Run make_gss_doc.R and make_gss_all_labelled.R first
 
 library(tidyverse)
 library(lubridate)
@@ -10,6 +10,9 @@ library(labelled)
 
 ## gss data
 load(here::here("data", "gss_all.rda"))
+
+## Labelled gss obj ()
+load(here("data-raw", "objects", "gss_all_labelled.rda"))
 
 
 # make a data dictionary
@@ -38,6 +41,28 @@ gss_dict <- gss_dict |>
   left_join(gss_years) |>
   select(-levels)
 
+## Add ballot table
+gss_which_ballots <- function(variable) {
+  gss_all_labelled |>
+    select(year, ballot, {{variable}}) |>
+    filter({{variable}} != "iap") |>
+    group_by(year, ballot) |>
+    tally() |>
+    pivot_wider(names_from = ballot, values_from = n) |>
+    mutate(across(starts_with("ballot"), \(x) replace_na(as.character(x), "-")))
+}
+
+gss_ballot_tbl <- tibble(
+  variable = gss_dict$variable
+)
+
+gss_ballot_tbl <- gss_ballot_tbl |>
+  mutate(var_ballots = map(variable, gss_which_ballots)) |>
+  mutate(var_ballots = map(var_ballots, as_tibble))
+
+
+gss_dict <- gss_dict |>
+  left_join(gss_ballot_tbl, by = "variable")
 
 ## Next we join gss_doc, created in make_gss_doc.R
 load(file = here::here("data", "gss_doc.rda"))
@@ -47,7 +72,8 @@ gss_dict <- left_join(gss_dict, gss_doc, by = "variable") |>
   relocate(var_doc_label, .before = value_labels) |>
   relocate(years, .before = var_yrtab) |>
   relocate(col_type:var_type, .after = var_yrtab) |>
-  relocate(var_na_codes, .after = var_type)
+  relocate(var_na_codes, .after = var_type) |>
+  relocate(var_ballots, .after = var_yrtab)
 
 ## Save out
 usethis::use_data(gss_dict, overwrite = TRUE, compress = "xz")